Bugfix: Detect custom (locally-compiled) Greyhole VFS module, and don't overwrite it

+++ greyhole.spec
@@ -121,8 +121,13 @@ if [ ! -f ${SOURCE_LIB} ]; then
 	exit 1
 fi
 
-rm -f ${TARGET_SYMLINK}
-ln -s ${SOURCE_LIB} ${TARGET_SYMLINK}
+ls -l "${TARGET_SYMLINK}" 2>/dev/null | grep '/usr/share/greyhole/vfs-build/' >/dev/null
+if [ $? -eq 0 ]; then
+	echo "Detected custom (locally-compiled) Greyhole VFS module at ${TARGET_SYMLINK}; will NOT overwrite it."
+else
+	rm -f ${TARGET_SYMLINK}
+	ln -s ${SOURCE_LIB} ${TARGET_SYMLINK}
+fi
 
 if [ -f /proc/fs/cifs/OplockEnabled ]; then
 	# cifs client workaround
